refactor: set naming according to google style, rm forward definition
